refactor: move EpicFightControlifyEntrypoint.BindContexts -> EpicFightControlifyBindContexts
